% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dySim-package.R
\docType{package}
\name{dySim-package}
\alias{dySim}
\alias{dySim-package}
\title{dySim: Dyadic Data Analysis Simulation}
\description{
Convenience functions for simulating data using a variety of popular models of dyadic data for both data generating mechanisms and the analytic model(s) applied within simulated sample.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsakaluk/dySim}
  \item Report bugs at \url{https://github.com/jsakaluk/dySim/issues}
}

}
\author{
\strong{Maintainer}: John Sakaluk \email{jksakaluk@gmail.com} (\href{https://orcid.org/0000-0002-2515-9822}{ORCID})

}
\keyword{internal}
