% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParams.R
\name{generateParams}
\alias{generateParams}
\alias{generateObsNames}
\alias{generateLVNames}
\alias{generatePopLoadsErrors}
\alias{generatePopCFMLoadsErrors}
\alias{generatePopResidCorrs}
\alias{generateICCs}
\alias{generateIPCs}
\alias{generateLVariances}
\alias{generateActors}
\alias{generatePartners}
\alias{generatePopCFMDySlopeError}
\title{Helper functions for scripters to generate parameter values of particular kinds of
parameters for dyadic SEMs using lavaan notation}
\usage{
generateObsNames(lv = "X", person = "A", items)

generateLVNames(lv = "X", person = "A")

generatePopLoadsErrors(var = "X", partner = "A", type = NULL, num = NULL)

generatePopCFMLoadsErrors(var = "X", type = NULL)

generatePopResidCorrs(var = "X", type = NULL, num = NULL)

generateICCs(var = "X", type = NULL)

generateIPCs(var = "A", type = NULL)

generateLVariances(var = "X", value = 1, partner = "A")

generateActors(partner = "A", type = NULL)

generatePartners(partner = "A", type = NULL)

generatePopCFMDySlopeError(type = NULL)
}
\arguments{
\item{lv}{character for name of LV}

\item{person}{character designation for a partner}

\item{items}{number of items for which parameters are needed}

\item{var}{character for stem of items for a given LV}

\item{partner}{character designation for a partner}

\item{type}{character input for magnitude threshold heuristic for "weak", "moderate", "strong", and "very strong" (for first or lower-order loadings)
and any combination therein (e.g., "weak_strong", "very strong_very strong") (for higher-order loadings as in CFM)}

\item{num}{number of items for which parameters are needed}

\item{value}{numeric value for a given parameter}
}
\value{
vector (or list of  vectors) of lavaan script elements and/or parameter values
}
\description{
Helper functions for scripters to generate parameter values of particular kinds of
parameters for dyadic SEMs using lavaan notation
}
\seealso{
Other helpers: 
\code{\link{getOutput}},
\code{\link{saveParamDf}},
\code{\link{scriptPopMods}}
}
\concept{helpers}
