% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\alias{getParams}
\title{Helper functions for getting output from sample lavaan models}
\usage{
getParams(fit, first = NULL, sampCount = NULL)
}
\arguments{
\item{fit}{lavaan object fit to a particular random sample's simulated data}

\item{first}{logical for whether output is from first random sample (TRUE) or not}

\item{sampCount}{counter of the current sample}
}
\value{
df of estimated parameter values and their statistical tests
}
\description{
Helper functions for getting output from sample lavaan models
}
\seealso{
Other helpers: 
\code{\link{generateParams}},
\code{\link{saveParamDf}},
\code{\link{scriptPopMods}}
}
\concept{helpers}
