% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDySim.R
\name{runDySim}
\alias{runDySim}
\title{Title}
\usage{
runDySim(
  seed = NULL,
  popMod = NULL,
  popModList = NULL,
  sampSize = NULL,
  sampMod = NULL,
  sampModType = NULL,
  nSims = NULL,
  output = "paramTable"
)
}
\arguments{
\item{seed}{numeric input for random seed to use}

\item{popMod}{character of dyadic model to use as population data generating mechanism. Currently supports "L-APIM".}

\item{popModList}{popModList with necessary parameter options for a given popMod}

\item{sampSize}{numeric input for sample size of each randomly drawn sample}

\item{sampMod}{character of dyadic model to use as analytic model within each randomly drawn sample. Currently supports "APIM"}

\item{sampModType}{character of type of specification for chosen sampMod. "latent" or "observed"}

\item{nSims}{numeric input for number of randomly drawn samples to draw from popMod}

\item{output}{character of type of output to extract from each analyzed sampMod. Currently supports "paramTable"}
}
\value{
a tibble of chosen output from nSims random samples
}
\description{
Title
}
\examples{
popModList.apim <- list(nItemsX = 5,
loadValuesX_A = "moderate", loadValuesX_B = "moderate",
residCorrValuesX = "moderate", iccX = "weak",
nItemsY = 5,
loadValuesY_A = "weak", loadValuesY_B = "mixed",
residCorrValuesY = "mixed", iccY = "strong",
actorA = "moderate", actorB = "very strong",
partnerA = "weak", partnerB = "weak")

apim.sim.out <- runDySim(seed = 123,
popMod = "L-APIM",
popModList = popModList.apim,
sampSize = 100,
sampMod = "APIM",
sampModType = "latent",
nSims = 10,
output = "paramTable")

}
