% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParamDf.R
\name{saveParamDf}
\alias{saveParamDf}
\alias{saveParamLAPIM}
\title{Helper functions for saving parameters randomly generated by scripPopMods
in order to compare simulated statistics against parameter values in tables and/or visualizations}
\usage{
saveParamLAPIM(paramList, paramType)
}
\arguments{
\item{paramList}{list of parameters generated by scriptPopMods}

\item{paramType}{character of what kind of parameters to save. Can be "structural", "measurement", or "both"}
}
\value{
data frame of parameter values styled to be rbind() friendly for output from runDySim (and \code{lavaan::parameterestimates} generally)
}
\description{
Helper functions for saving parameters randomly generated by scripPopMods
in order to compare simulated statistics against parameter values in tables and/or visualizations
}
\seealso{
Other helpers: 
\code{\link{generateParams}},
\code{\link{getOutput}},
\code{\link{scriptPopMods}}
}
\concept{helpers}
