% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptPopMods.R
\name{scriptPopMods}
\alias{scriptPopMods}
\alias{scriptPopModAPIM}
\alias{scriptPopModCFM}
\title{Helper functions for taking output from generateParams helpers
to write population model script using lavaan notation}
\usage{
scriptPopModAPIM(popModList, ...)

scriptPopModCFM(popModList)
}
\arguments{
\item{popModList}{list with necessary parameter options for a given popMod}

\item{...}{arguments to pass onto helper functions (e.g., saveParamDf)}
}
\value{
a lavaan script with parameter values for a latent APIM
}
\description{
Helper functions for taking output from generateParams helpers
to write population model script using lavaan notation
}
\examples{
popModList.apim <- list(nItemsX = 5,
loadValuesX_A = "moderate", loadValuesX_B = "moderate",
residCorrValuesX = "moderate", iccX = "weak",
nItemsY = 5,
loadValuesY_A = "weak", loadValuesY_B = "mixed",
residCorrValuesY = "mixed", iccY = "strong",
actorA = "moderate", actorB = "very strong",
partnerA = "weak", partnerB = "weak")

scriptPopModAPIM(popModList = popModList.apim, paramType = "structural")
}
\seealso{
Other helpers: 
\code{\link{generateParams}},
\code{\link{saveParamDf}}
}
\concept{helpers}
